+++ src/Plugin.cs
@@ -2,15 +2,21 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
+using System.Runtime.Serialization.Formatters.Binary;
 using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
+using GameNetcodeStuff;
 using HarmonyLib;
+using LethalLib;
 using UnityEngine;
 using LethalLib.Modules;
+using Unity.Collections;
+using Unity.Netcode;
 using static LethalLib.Modules.Levels;
 using static LethalLib.Modules.Enemies;
 using static LethalLib.Modules.Items;
+using NetworkPrefabs = LethalLib.Modules.NetworkPrefabs;
 
 namespace LethalCompanyHarpGhost
 {
@@ -23,7 +29,7 @@ namespace LethalCompanyHarpGhost
 
         private readonly Harmony _harmony = new Harmony(ModGuid);
         
-        private static ManualLogSource _mls;
+        private static readonly ManualLogSource Mls = BepInEx.Logging.Logger.CreateLogSource(ModGuid);
 
         private static HarpGhostPlugin _instance;
         
@@ -36,17 +42,16 @@ namespace LethalCompanyHarpGhost
         private void Awake()
         {
             if (_instance == null) _instance = this;
-            
-            _mls = BepInEx.Logging.Logger.CreateLogSource(ModGuid);
 
             Assets.PopulateAssets();
             if (Assets.MainAssetBundle == null)
             {
-                _mls.LogError("MainAssetBundle is null");
+                Mls.LogError("MainAssetBundle is null");
                 return;
             }
 
-            // HarpGhostConfig = new HarpGhostConfig(Config);
+            HarpGhostConfig = new HarpGhostConfig(Config);
+            _harmony.PatchAll(typeof(HarpGhostConfig));
             
             SetupHarpGhost();
             SetupHarp();
@@ -68,7 +73,7 @@ namespace LethalCompanyHarpGhost
                 }
             }
             
-            _mls.LogInfo($"Plugin {ModName} is loaded!");
+            Mls.LogInfo($"Plugin {ModName} is loaded!");
         }
 
         private static void SetupHarpGhost()
@@ -79,7 +84,10 @@ namespace LethalCompanyHarpGhost
             
             NetworkPrefabs.RegisterNetworkPrefab(_harpGhost.enemyPrefab);
             Utilities.FixMixerGroups(_harpGhost.enemyPrefab);
-            RegisterEnemy(_harpGhost, SpawnType.Default, new Dictionary<LevelTypes, int>{[LevelTypes.DineLevel] = 40, [LevelTypes.RendLevel] = 40}, infoNode: harpGhostTerminalNode, infoKeyword:harpGhostTerminalKeyword);
+            RegisterEnemy(_harpGhost, SpawnType.Default, new Dictionary<LevelTypes, int>{
+                [LevelTypes.DineLevel] = HarpGhostConfig.Instance.GhostSpawnRate.Value, 
+                [LevelTypes.RendLevel] = HarpGhostConfig.Instance.GhostSpawnRate.Value}, 
+                infoNode: harpGhostTerminalNode, infoKeyword:harpGhostTerminalKeyword);
         }
 
         private static void SetupHarp()
@@ -93,7 +101,7 @@ namespace LethalCompanyHarpGhost
             HarpItem = Assets.MainAssetBundle.LoadAsset<Item>("HarpItemData");
             if (HarpItem == null)
             {
-                _mls.LogError("Failed to load HarpItemData from AssetBundle.");
+                Mls.LogError("Failed to load HarpItemData from AssetBundle.");
                 return;
             }
             
@@ -103,60 +111,223 @@ namespace LethalCompanyHarpGhost
         }
     }
 
-    public class HarpGhostConfig
+    public class HarpGhostConfig : SyncedInstance<HarpGhostConfig>
     {
-        public static ConfigEntry<int> ConfigInitialHealth;
-        public static ConfigEntry<int> ConfigAttackDamage;
+        public readonly ConfigEntry<int> GhostInitialHealth;
+        public readonly ConfigEntry<int> GhostAttackDamage;
         
-        public static ConfigEntry<float> ConfigAnnoyanceLevelDecayRate;
-        public static ConfigEntry<float> ConfigAnnoyanceThreshold;
-        public static ConfigEntry<float> ConfigMaxSearchRadius;
+        public readonly ConfigEntry<float> GhostAnnoyanceLevelDecayRate;
+        public readonly ConfigEntry<float> GhostAnnoyanceThreshold;
+        public readonly ConfigEntry<float> GhostMaxSearchRadius;
 
-        public static ConfigEntry<float> ConfigVoiceSfxVolume;
+        public readonly ConfigEntry<float> GhostVoiceSfxVolume;
+        public readonly ConfigEntry<float> HarpMusicVolume;
+
+        public readonly ConfigEntry<int> GhostSpawnRate;
+        public readonly ConfigEntry<int> HarpMinValue;
+        public readonly ConfigEntry<int> HarpMaxValue;
 
         public HarpGhostConfig(ConfigFile cfg)
         {
-            ConfigInitialHealth = cfg.Bind(
+            InitInstance(this);
+            
+            GhostInitialHealth = cfg.Bind(
                 "General",
                 "Health",
                 3,
                 "The health when spawned"
                 );
             
-            ConfigAttackDamage = cfg.Bind(
+            GhostAttackDamage = cfg.Bind(
                 "General",
                 "Attack Damage",
                 35,
                 "The attack damage of the ghost"
             );
             
-            ConfigAnnoyanceLevelDecayRate = cfg.Bind(
+            GhostAnnoyanceLevelDecayRate = cfg.Bind(
                 "General",
                 "Annoyance Level Decay Rate",
                 0.3f,
                 "The decay rate of the ghost's annoyance level (due to noises) over time"
             );
             
-            ConfigAnnoyanceThreshold = cfg.Bind(
+            GhostAnnoyanceThreshold = cfg.Bind(
                 "General",
                 "Annoyance Level Threshold",
                 8f,
                 "The threshold of how annoyed the ghost has to be (from noise) to get angry"
             );
             
-            ConfigMaxSearchRadius = cfg.Bind(
+            GhostMaxSearchRadius = cfg.Bind(
                 "General",
                 "Max Search Radius",
                 100f,
                 "The maximum distance the ghost will go to search for a player"
             );
             
-            ConfigVoiceSfxVolume= cfg.Bind(
-                "General",
+            GhostVoiceSfxVolume = cfg.Bind(
+                "Audio",
                 "Voice Sound Effects Volume",
                 1f,
                 "The volume of the ghost's voice. Values are from 0-1"
             );
+
+            HarpMusicVolume = cfg.Bind(
+                "Audio",
+                "Harp Music Volume",
+                1f,
+                "The volume of the music played from the harp. Values are from 0-1"
+            );
+            
+            GhostSpawnRate = cfg.Bind(
+                "Spawn Values",
+                "Harp Ghost Spawn Rates",
+                40,
+                "The weighted spawn rarity of the harp ghost"
+            );
+            
+            HarpMinValue = cfg.Bind(
+                "Spawn Values",
+                "Harp Minimum value",
+                150,
+                "The minimum value that the harp can be set to"
+            );
+            
+            HarpMaxValue = cfg.Bind(
+                "Spawn Values",
+                "Harp Maximum value",
+                300,
+                "The maximum value that the harp can be set to"
+            );
+        }
+
+        private static void RequestSync() {
+            if (!IsClient) return;
+
+            using FastBufferWriter stream = new(IntSize, Allocator.Temp);
+            MessageManager.SendNamedMessage("ModName_OnRequestConfigSync", 0uL, stream);
+        }
+
+        private static void OnRequestSync(ulong clientId, FastBufferReader _) {
+            if (!IsHost) return;
+
+            Debug.Log($"Config sync request received from client: {clientId}");
+
+            byte[] array = SerializeToBytes(Instance);
+            int value = array.Length;
+
+            using FastBufferWriter stream = new(value + IntSize, Allocator.Temp);
+
+            try {
+                stream.WriteValueSafe(in value, default);
+                stream.WriteBytesSafe(array);
+
+                MessageManager.SendNamedMessage("ModName_OnReceiveConfigSync", clientId, stream);
+            } catch(Exception e) {
+                Debug.Log($"Error occurred syncing config with client: {clientId}\n{e}");
+            }
+        }
+
+        private static void OnReceiveSync(ulong _, FastBufferReader reader) {
+            if (!reader.TryBeginRead(IntSize)) {
+                Debug.LogError("Config sync error: Could not begin reading buffer.");
+                return;
+            }
+
+            reader.ReadValueSafe(out int val, default);
+            if (!reader.TryBeginRead(val)) {
+                Debug.LogError("Config sync error: Host could not sync.");
+                return;
+            }
+
+            byte[] data = new byte[val];
+            reader.ReadBytesSafe(ref data, val);
+
+            SyncInstance(data);
+
+            Debug.Log("Successfully synced config with host.");
+        }
+        
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(PlayerControllerB), "ConnectClientToPlayerObject")]
+        public static void InitializeLocalPlayer() {
+            if (IsHost) {
+                MessageManager.RegisterNamedMessageHandler("HarpGhost_OnRequestConfigSync", OnRequestSync);
+                Synced = true;
+
+                return;
+            }
+
+            Synced = false;
+            MessageManager.RegisterNamedMessageHandler("HarpGhost_OnReceiveConfigSync", OnReceiveSync);
+            RequestSync();
+        }
+        
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(GameNetworkManager), "StartDisconnect")]
+        public static void PlayerLeave() {
+            RevertSync();
+        }
+    }
+
+    // Got from https://lethal.wiki/dev/intermediate/custom-config-syncing
+    [Serializable]
+    public class SyncedInstance<T>
+    {
+        internal static CustomMessagingManager MessageManager => NetworkManager.Singleton.CustomMessagingManager;
+        internal static bool IsClient => NetworkManager.Singleton.IsClient;
+        internal static bool IsHost => NetworkManager.Singleton.IsHost;
+        
+        [NonSerialized]
+        protected static int IntSize = 4;
+
+        public static T Default { get; private set; }
+        public static T Instance { get; private set; }
+
+        public static bool Synced { get; internal set; }
+
+        protected void InitInstance(T instance) {
+            Default = instance;
+            Instance = instance;
+            
+            IntSize = sizeof(int);
+        }
+
+        internal static void SyncInstance(byte[] data) {
+            Instance = DeserializeFromBytes(data);
+            Synced = true;
+        }
+
+        internal static void RevertSync() {
+            Instance = Default;
+            Synced = false;
+        }
+
+        public static byte[] SerializeToBytes(T val) {
+            BinaryFormatter bf = new();
+            using MemoryStream stream = new();
+
+            try {
+                bf.Serialize(stream, val);
+                return stream.ToArray();
+            }
+            catch (Exception e) {
+                Debug.LogError($"Error serializing instance: {e}");
+                return null;
+            }
+        }
+
+        public static T DeserializeFromBytes(byte[] data) {
+            BinaryFormatter bf = new();
+            using MemoryStream stream = new(data);
+
+            try {
+                return (T) bf.Deserialize(stream);
+            } catch (Exception e) {
+                Debug.LogError($"Error deserializing instance: {e}");
+                return default;
+            }
         }
     }