+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0 Out of Beta
+* Configs added, more will be added upon request 
+* Better scrap icon (one that isn't a white square)
+
 ## 1.1.3
 * Two more harp songs added
 * Shotgun now does damage